Require Send + Sync for object types
Without this, it is possible to send types that contain non-threadsafe
constructs like `RefCell` over FFI, and call methods that use them from
multiple threads.

Both Send and Sync are required on the inner type for Arc to implement
Send itself, which is what we must guarantee to be able to move an Arc
over FFI, since we can't guarantee it is kept on the same thread inside
the calling code.