Added ByRef support for proc-macros
The nice thing about this change is that users don't need to mark their
arguments in some special way.  If you take a `&Object` things will
work.

The not-so-nice parts is that we need to go back to token parsing to
implement this, so type aliases can make it fail.  Also it means that we
can't support things like `&Enum` or `&Record`.  I thought about trying
some tricks like `TYPE_ID_META` or introducing a new trait, but I
couldn't think of any way to make it work.  Given that users used to
have to mark the argument with `[ByRef]` and this hasn't historically
worked with enums/records, I think it should be okay.