Allow procmacros to interact with UDL when namespace != crate_name (#1728)
`uniffi_bindgen::generate_scaffolding` etc now attempt to locate and parse
the corresponding Cargo.toml. New functions, typically with a `_for_crate`
suffix have been made available which take the crate_name as an arg,
allowing this .toml parsing to be skipped.

The name of the various FFI functions are now always based on this crate
name instead of the namespace name.

The UDL_FILE metadata value written to the scaffolding is now created by
scaffolding templates rather than procmacros so it can carry more info about
how the scaffolding was generated - in particular, the crate_name, so
sanity checks can be done as metadata is loaded from the built .so.