Reworked proc macro argument lifting / passing
Split scaffolding calls into two phases:
  - lift: Try to lift all arguments.
  - rust call: Call the Rust function or handle the lift error.

This fixes 2 bugs with async functions:
  - They couldn't input objects
  - They couldn't handle argument conversion errors.

The reason for this split that async Futures need to be Send, but the
lift phase isn't always Send because it may deal with raw pointers.  The
new code performs the lift outside of the future, then the future deals
with the result of the lift. This means that the lift phase is executed
outside of the async context, but it should be very fast.

Changed `handle_failed_lift` signature. Now, if a `Result<>` is able to
downcast the error, it returns `Self::Err` directly instead of
serializing the error into `RustBuffer`.