Swift cancellation
Added a way for users to opt-in to async cancellation checks on Swift.
This means that if the parent Task is cancelled, the exported Rust
function will raise a `CancellationError`.  It's opt-in because this
means that the function will always be marked `throws` forcing users to
add `try` to all calls.

Reworked the `rustCall*` functions a bit:
  * Split `uniffiRustCallAsync` into a cancellable and non-cancellable
    version.
  * Added the `uniffi` prefix `rustCall` and `rustCallWithError`.
  * `uniffiRustCall` is no longer marked `throws`, since it's not
    expected to throw.  `uniffiRustCallWithError` is still marked
    throws.
  * `uniffiRustCallAsync` now inputs a complete closure that calls the
    FFI complete function with either `uniffiRustCall` or
    `uniffiRustCallWithError`
  * The result of this is that we can now use `rethrows` and Swift will
    figure out if an async function will throw or not.