Use Handle for passing foreign objects across the FFI
Exported Rust FFI functions to manage a Slab that stores unit values.
The foreign code uses this to allocate and manage handles then stores the
actual data inside a native array.

This replaces the previous handle maps / pointer leaking code for
callbacks, trait interfaces and continuation data.

Started updating the ForeignExecutor code to use handles, but this is
still a WIP while the ForeignExecutor type is in it's limbo state.