Use handles for better trait interfaces passing
Check if the trait interface handle originates from the other side of
the FFI.  If it does, inc-ref it and return the handle, rather than
returning a handle to the wrapped object.  Before, each time the trait
object was passed across the FFI, we wrapped it another time

On the foreign side, this can be accomplished by a type check.  On the
Rust side, this requires an extra, `#[doc(hidden)]` method on the trait.
This means that UDL-defined trait interfaces need to be wrapped with an
attribute macro that inserts that method.

One issue with this is that it can cause us to leak references if we do
the inc-ref, then there's an exception lowering another argument.  I
believe this can be fixed by separating out the failable lowering code
from the non-failable code.