Use handles for better trait interfaces passing
Check if the trait interface handle originates from the other side of
the FFI.  If it does, clone the handle, rather than returning a handle
to the wrapped object.  Before, each time the trait object was passed
across the FFI, we wrapped it another time

On the foreign side, this can be accomplished by a type check.  On the
Rust side, this requires an extra, `#[doc(hidden)]` method on the trait.
This means that UDL-defined trait interfaces need to be wrapped with an
attribute macro that inserts that method.

Reworked the changelog so that breaking changes for external bindings is
its own section.

Removed the reexport-scaffolding-macro fixture which often requires
changes when the FFI changes.  I don't think it's giving us enough value
at this point to justify continuing to update it.