Make foreign trait implementations opt-in (#1827)
This changes the default behavior for exported trait interfaces to not
generate a foreign callback interface implementation for them.  Users
can opt-in to this using the `WithForeign` attribute for UDL and the
`with_foreign` for proc-macros.

The main reason for this change is that callback interfaces and trait
interfaces aren't completely compatible.  For example, `ByRef` is not
supported for callback interfaces, but it is supported for trait
interfaces that only have Rust implementations.