Implement LiftRef for str and [T]
This is needed to allow these to work with proc-macros.

This is not needed for UDL, because there we generate a function than
inputs `&String`, then pass that value to the Rust function than inputs
`&str`.  Rust can handle that conversion automatically.  However, with
proc-macros the scaffolding function types are always exactly the same
as the source function types.