Be more consistent with Python variable prefixes
Always use `_uniffi`, `_Uniffi`, or `_UNIFFI` for module-level names.
The leading underscore is to indicate private variables and makes it so
the names don't appear in autocomplete. I think this is the most
Pythonic naming style.

Use `_uniffi` rather than simply `_` to avoid name collisions. I think
the general contract with our users should be that we own the `uniffi`
prefix.  If they want to name something `UniffiFoo`, then they take the
risk of a name collision.