Add support for non-exhaustive enums (#1593)
UDL types can now have the `[NonExhaustive]` attribute.  If present, we
add a default branch arm to make things compile correctly with
non-exhaustive enums.  I didn't want to make this the default behavior,
since in the normal case we want missing variants to result in compile
errors.

In theory, this also adds the `non_exhaustive` attribute for
proc-macros.  However, there's no way to test this because
`non_exhaustive` only matters for types defined in remote crates and
there's currently no proc-macro support for those.

One slightly silly part of this is the flow of metadata.  It doesn't
make much sense as a metadata field, since it doesn't affect bindings
generation at all.  However, it needs to be because of the way that data
flows from `uniffi_udl` -> `uniffi_meta` -> `uniffi_bindgen` ->
`uniffi_macros`.