Be more consistent with Kotlin variable prefixes (#1907)
Let's always use `uniffi`, `Uniffi`, or `UNIFFI`. Before some of the
variable names had leading underscores, which seems redundant and not
idomatic Kotlin.  My guess is that it was copy and pasted from Python.

I think the general contract with our users should be that we own the
`uniffi` prefix.  If they want to name something `UniffiFoo`, then they
take the risk of a name collision.