Remove uniffi_meta::ErrorMetadata. (#1936)
This means that we determine whether a name is used as an error by checking if,
well, it's used as an error - it's not impacted by how the name is declared. This has
been done primarily for future "objects as errors" work where we don't want to force
an object to be declared specially to have it be able to be used as an error.

This has a side-effect which should be invisible to most users but is visible in our
test suites - some of our test UDL declares an enum with `[Error]` but doesn't actually
use that enum as an error. Previously this enum would get error semantics, whereas now
it has been necessary to change these fixtures to use the object as an error. This should
be invisible to real users because objects declared as being an error are either (a) used
as an error or (b) unused, where the actual semantics don't matter.