Allow callback interfaces to return other callback interfaces (#1947)
The easy part was implementing `LiftReturn`.  The hard part was figuring
out how to handle TYPE_ID_META in the metadata code.  In general, we
have an issue: should we get that from `Lift`/`LiftReturn` or
`Lower`/`LowerReturn`.

As mentioned in the `fnsig` code, I feel like this is a bid of a
band-aid.  However, I think a real fix would require significant changes
to the FFI trait system and I don't want to do that so close to the
v0.26 release.