Fix Kotlin inheritance loop for `Error`
If a Rust error enum is named `FooError`, it becomes `FooException` in
the generated Kotlin. The `FooException` class inherits from
`Exception`, which is a built-in Kotlin class.

If the Rust error enum is instead called `Error`, the generated
`Exception` Kotlin class fails to disambiguate between itself and the
built-in Kotlin class.

This can be fixed by qualifying the class as `kotlin.Exception` to
unambiguously refer to this class:
https://kotlinlang.org/api/latest/jvm/stdlib/kotlin/-exception/