Update toml dependency to latest version
I am in the process of vendoring UniFFI in AOSP, which means that I
have to make it play nice with the existing crates found here:

  https://cs.android.com/android/platform/superproject/+/main:external/rust/crates/

Android tries to only vendor each crate in a single version, so it
would be helpful for me if the toml dependency could be updated.

This updates the “Requires” version to 1.67. I believe this is okay
since this version is below 1.74 which is what mozilla-central
currently requires.