Combine ExportedImplFnArgs and ExportFnArgs
These were basically the same, so using 1 struct simplifies the code.

The one difference was that ExportFnArgs allowed an async runtime to be
specified.  Now that the 2 are merged, it's possible to specify an async
runtime for constructors/methods.  If present that runtime will override
the one specified in the impl attribute.