Remove uniffi::generate_component_scaffolding
It's a direct re-export of uniffi_bindgen::generate_component_scaffolding and
behind the 'bindgen' feature. For this reason it's not actually used in
build scripts - they use 'uniffi::generate_scaffolding' which is correctly behind the
'build' feature.

We can't commit the "bindgen" feature to generating scaffolding, and it's kinda confusing, so let's kill it.