Adding #[remote] attribute
This can be used from proc-macros to specify that an item is remote and
therefore the FFI traits should only be implemnted for the local
`UniFfiTag`.

Added an example where this is used to use anyhow::Error in as your
error type.

Also added the `[Remote]` tag for UDL.  It doesn't do anything yes
because all UDL-based generation currently uses the local tag. I'm going
to change that in the next commit though.

Require that error types to implement `Display + Debug` rather than `Error`.
For some reason `anyhow::Error` doesn't actually implement error.