Adding the `LowerError` trait
This specifies how types are lowered when they are the `E` side of
`Result<T, E>`.  The main reason for the trait is that it lets us
specialize the handling for interfaces by auto-wrapping the value in an
Arc.

This makes using interfaces as errors much more ergonomic.

Also, replace the `Error` bound on the `E` types with `Display + Debug`.
I just found out that `anyhow::Error` doesn't actually implement `Error`
for some obscure reason.