Switch to `test --no-run` from `build` for cdylib
`uniffi-bindgen-java` is external to the uniffi repo, so the
fixtures/examples are all `dev-dependencies`, which aren't built on a
call to `cargo build`. `cargo test --no-run` causes them to be built but
doesn't cause a run of tests in place.

`.no_deps()` seems like it has a similar reaction as an external
bindings generator, it ignores dependencies in the cargo metadata, which
is where all the fixtures/examples used in tests cdylibs will be.