Fix so cargo_metadata isn't a dependency without the cargo_metadata feature. (#2200)
The `uniffi_test` crate unconditionally depends on `cargo_metadata`, and this will be
difficult to change. This means that because `uniffi_test` is an unconditional dependency
of `uniffi_bindgen`, the `cargo_metadata` feature doesn't actually prevent `uniffi_bindgen`
from depending on `cargo_metadata`.

The top level `uniffi` create already has a feature `bindgen-tests`. This patch
introduces a feature of the same name to `uniffi_bindgen`. Enabling that feature also
enables the metadata feature. This means that if you do not enable either of these features
you should be able to avoid `cargo_metadata`.

This patch also renames the new `cargo_metadata` feature to be `cargo-metadata` to
be more consistent with our existing features and Rust feature naming in general.