Provide more external type information to `FfiType::RustBuffer` (#2195)
* Thread `ExternalType` metadata into RustBuffer

In service of Java bindgen being able to generate fully qualified
`RustBuffer`s when necessary.

* Switch to `test --no-run` from `build` for cdylib

`uniffi-bindgen-java` is external to the uniffi repo, so the
fixtures/examples are all `dev-dependencies`, which aren't built on a
call to `cargo build`. `cargo test --no-run` causes them to be built but
doesn't cause a run of tests in place.