Expose the find_components function
The generate_bindings docs mentioned wanting to expose a function that
just finds ComponentInterface and config tables for each component and
leaves the rest to the external bindings generator.  This is exactly
what I want to use for uniffi-bindgen-gecko-js.

The `find_components` function was pretty much exactly what we wanted.
I made some minor changes to it and exposed it as a pub function.