Hide internal classes generated docs
Some of these could be made internal, but most could not.  The reason is
that UniFFI code needs to import them from other modules in order to
support external types.  This change mostly adds docstrings/annotations
to hide them from the generated docs.

For swift, this needed to behind an `#if` statement to keep
compatability with 5.5.  Maybe we can up the minimum version with the
next breaking release.