Switch Python to BindingsIr
The new system is to do a `BindingsIr` specialization pass, which means
implementing `VisitMut` and using it to:
  - Rewrite names, docstrings, etc.
  - Derive things like FFI converter names and store them in `lang_data`
  - Generate language-specific things like imports as we walk the tree

Along the way I made a couple changes to the Python bindings:

Changed `_uniffi_rust_call_async` to not lift the return value.  This
makes it fit in with the sync logic better and also enables async
constructors (which store the pointer in `self` rather than lift it).

Added the `CustomTypeConfig::type_name`.  While updating the code I
noticed the current templates handle this in a slightly buggy way where
they assume the type name is the builtin type, even if there's a
`from_custom`/`into_custom` that converts to a different type.

Removed almost all of the macros, `define_callable` is the only one left.